+++ src/structs.hpp
@@ -5,13 +5,13 @@
 struct NetworkConfiguration
 {
   std::string model_name;
-  std::vector<size_t> &hidden_layers;
+  std::vector<int> &hidden_layers;
   float &learning_rate;
-  size_t &batch_size;
+  int &batch_size;
 };
 
 struct NetworkInitialization {
-    std::vector<size_t> &layers_size;
+    std::vector<int> &layers_size;
     std::vector<Mann::Matrix> &nodes;
     std::vector<Mann::Matrix> &weights;
     std::vector<Mann::Matrix> &biases;
@@ -19,5 +19,5 @@ struct NetworkInitialization {
 
 struct NetworkArchitecture {
     NetworkInitialization* network_initialization;
-    std::vector<size_t> &hidden_layers_size;
+    std::vector<int> &hidden_layers_size;
 };
